# Performance improvements to the RewindDataIndexCache