--- a/src/Stratis.Bitcoin.Features.Consensus.Tests/CoinViews/RewindDataIndexCacheTest.cs
+++ b/src/Stratis.Bitcoin.Features.Consensus.Tests/CoinViews/RewindDataIndexCacheTest.cs
@@ -38,7 +38,7 @@ namespace Stratis.Bitcoin.Features.Consensus.Tests.CoinViews
 
             await rewindDataIndexCache.InitializeAsync(5, coinViewMock.Object);
 
-            var items = rewindDataIndexCache.GetMemberValue("items") as ConcurrentDictionary<string, int>;
+            var items = rewindDataIndexCache.GetMemberValue("items") as Dictionary<RewindDataIndexItem, int>;
 
             items.Should().HaveCount(10);
             this.CheckCache(items, 5, 1);
@@ -55,7 +55,7 @@ namespace Stratis.Bitcoin.Features.Consensus.Tests.CoinViews
 
             await rewindDataIndexCache.InitializeAsync(20, coinViewMock.Object);
 
-            var items = rewindDataIndexCache.GetMemberValue("items") as ConcurrentDictionary<string, int>;
+            var items = rewindDataIndexCache.GetMemberValue("items") as Dictionary<RewindDataIndexItem, int>;
 
             items.Should().HaveCount(22);
             this.CheckCache(items, 20, 10);
@@ -73,7 +73,7 @@ namespace Stratis.Bitcoin.Features.Consensus.Tests.CoinViews
             await rewindDataIndexCache.InitializeAsync(20, coinViewMock.Object);
 
             rewindDataIndexCache.Save(new Dictionary<RewindDataIndexItem, int>() { { new RewindDataIndexItem(new uint256(21), 0), 21}});
-            var items = rewindDataIndexCache.GetMemberValue("items") as ConcurrentDictionary<string, int>;
+            var items = rewindDataIndexCache.GetMemberValue("items") as Dictionary<RewindDataIndexItem, int>;
 
             items.Should().HaveCount(23);
             this.CheckCache(items, 21, 10);
@@ -91,7 +91,7 @@ namespace Stratis.Bitcoin.Features.Consensus.Tests.CoinViews
             await rewindDataIndexCache.InitializeAsync(20, coinViewMock.Object);
 
             rewindDataIndexCache.Flush(15);
-            var items = rewindDataIndexCache.GetMemberValue("items") as ConcurrentDictionary<string, int>;
+            var items = rewindDataIndexCache.GetMemberValue("items") as Dictionary<RewindDataIndexItem, int>;
 
             items.Should().HaveCount(12);
             this.CheckCache(items, 15, 9);
@@ -109,16 +109,16 @@ namespace Stratis.Bitcoin.Features.Consensus.Tests.CoinViews
             await rewindDataIndexCache.InitializeAsync(20, coinViewMock.Object);
 
             await rewindDataIndexCache.Remove(19, coinViewMock.Object);
-            var items = rewindDataIndexCache.GetMemberValue("items") as ConcurrentDictionary<string, int>;
+            var items = rewindDataIndexCache.GetMemberValue("items") as Dictionary<RewindDataIndexItem, int>;
 
             items.Should().HaveCount(22);
             this.CheckCache(items, 19, 9);
         }
 
 
-        private void CheckCache(ConcurrentDictionary<string, int> items, int tip, int bottom)
+        private void CheckCache(Dictionary<RewindDataIndexItem, int> items, int tip, int bottom)
         {
-            foreach (KeyValuePair<string, int> keyValuePair in items)
+            foreach (KeyValuePair<RewindDataIndexItem, int> keyValuePair in items)
             {
                 Assert.True(keyValuePair.Value <= tip && keyValuePair.Value >= bottom);
             }