Primary motivation behind this change is a performance improvement as outlined below:


|                                 Method | ArraySize | LookupCount |            Mean |           Error |          StdDev | Ratio | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------- |---------- |------------ |----------------:|----------------:|----------------:|------:|-----:|------------:|------------:|------------:|--------------------:|
|     **DictionaryWithStringKeyAdd_Current** |     **10000** |          **50** | **8,134,612.21 ns** | **135,792.3472 ns** | **127,020.2458 ns** |  **1.00** |    **1** |   **1562.5000** |    **578.1250** |    **281.2500** |           **8782030 B** |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
|    DictionaryWithObjectKeyAdd_Proposed |     10000 |          50 |   506,217.26 ns |  10,072.2140 ns |   9,421.5552 ns |  1.00 |    1 |    210.9375 |    160.1563 |    159.1797 |            942259 B |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
|   DictionaryWithStringKeyExist_Current |     10000 |          50 |       375.49 ns |       2.6040 ns |       2.4358 ns |  1.00 |    1 |      0.1864 |           - |           - |               784 B |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
|  DictionaryWithObjectKeyExist_Proposed |     10000 |          50 |        20.59 ns |       0.1060 ns |       0.0991 ns |  1.00 |    1 |           - |           - |           - |                   - |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
|  DictionaryWithStringKeyUpdate_Current |     10000 |          50 |    19,559.08 ns |      49.9271 ns |      41.6914 ns |  1.00 |    1 |      9.3384 |           - |           - |             39200 B |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
| DictionaryWithObjectKeyUpdate_Proposed |     10000 |          50 |     1,429.46 ns |       5.1146 ns |       4.7842 ns |  1.00 |    1 |           - |           - |           - |                   - |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
|  DictionaryWithStringKeyRemove_Current |     10000 |          50 |   733,154.18 ns |   6,026.9656 ns |   5,342.7481 ns |  1.00 |    1 |     76.1719 |     76.1719 |     76.1719 |            322216 B |
|                                        |           |             |                 |                 |                 |       |      |             |             |             |                     |
| DictionaryWithObjectKeyRemove_Proposed |     10000 |          50 |   368,923.53 ns |   1,661.2000 ns |   1,387.1772 ns |  1.00 |    1 |     76.6602 |     76.6602 |     76.6602 |            283016 B |