@dangershony 2019-03-26
Closing as the main improvement in this PR moved here 
https://github.com/stratisproject/StratisBitcoinFullNode/pull/3437