@MithrilMan 2019-03-26
just a generic comment about changing for loops from int to uint:
https://stackoverflow.com/a/35142888/1716620