@MithrilMan 2019-03-27
@maciejzaleski can you post your benchmark code?